fix: update validator comment and extract parseGqlError to shared utility
Agent-Logs-Url: https://github.com/mateo-khalil/sumate-ya/sessions/e2104dd1-8f42-4f6d-af54-85acce32b3f2
