+++ apps/frontend/src/components/matches/MatchResultsSection.tsx
@@ -16,9 +16,9 @@
  * - Styling: Tailwind utility classes using @theme tokens from globals.css.
  *   font-display → Bebas Neue, font-condensed → Barlow Condensed (tokens added for P3).
  *   bg-success / text-success-foreground → hsl(142 72% 50%) / hsl(142 72% 65%).
- * - Error display: vote errors from GraphQL are passed through parseGqlError so Zod v4 JSON
- *   arrays render as readable Spanish messages rather than raw JSON. Keep in sync with the
- *   same helper in ProposeResultForm.
+ * - Error display: vote errors from GraphQL are normalised via the shared parseGqlError
+ *   helper (apps/frontend/src/lib/parseGqlError.ts) so Zod v4 JSON arrays render as
+ *   readable Spanish messages rather than raw JSON.
  * - Previously fixed bugs:
  *   - P2 audit: "Cargar resultado" disappeared when all submissions were REJECTED.
  *   - P3 audit: inline style objects replaced with Tailwind classes per design-system.md.
@@ -33,30 +33,13 @@ import {
   type VoteValue,
 } from '../../graphql/operations/match-results.js';
 import ProposeResultForm from './ProposeResultForm.js';
+import { parseGqlError } from '../../lib/parseGqlError.js';
 
 interface Props {
   matchId: string;
   isParticipant: boolean;
   backendUrl: string;
   accessToken: string;
-}
-
-function parseGqlError(message: string): string {
-  try {
-    const parsed: unknown = JSON.parse(message);
-    if (
-      Array.isArray(parsed) &&
-      parsed.length > 0 &&
-      typeof (parsed[0] as Record<string, unknown>).message === 'string'
-    ) {
-      return parsed
-        .map((issue: Record<string, unknown>) => String(issue.message))
-        .join('. ');
-    }
-  } catch {
-    // not JSON — fall through
-  }
-  return message;
 }
 
 async function gql<T>(
